+++ example/src/screens/modal/SimpleExample.tsx
@@ -1,6 +1,9 @@
+import {
+  type BottomSheetHandleProps,
+  BottomSheetModal,
+} from '@gorhom/bottom-sheet';
 import React, { useCallback, useMemo, useRef, useState } from 'react';
-import { View, StyleSheet } from 'react-native';
-import { BottomSheetHandleProps, BottomSheetModal } from '@gorhom/bottom-sheet';
+import { StyleSheet, View } from 'react-native';
 import { Button } from '../../components/button';
 import { ContactList } from '../../components/contactList';
 import { HeaderHandle } from '../../components/headerHandle';
@@ -20,27 +23,25 @@ export const SimpleExample = () => {
 
   //#region callbacks
   const handleChange = useCallback((index: number) => {
-    // eslint-disable-next-line no-console
     console.log('index', index);
   }, []);
   const handleDismiss = useCallback(() => {
-    // eslint-disable-next-line no-console
     console.log('on dismiss');
   }, []);
   const handleDismissPress = useCallback(() => {
-    bottomSheetRef.current!.dismiss();
+    bottomSheetRef.current?.dismiss();
   }, []);
   const handleClosePress = useCallback(() => {
-    bottomSheetRef.current!.close();
+    bottomSheetRef.current?.close();
   }, []);
   const handleExpandPress = useCallback(() => {
-    bottomSheetRef.current!.expand();
+    bottomSheetRef.current?.expand();
   }, []);
   const handleCollapsePress = useCallback(() => {
-    bottomSheetRef.current!.collapse();
+    bottomSheetRef.current?.collapse();
   }, []);
   const handlePresentPress = useCallback(() => {
-    bottomSheetRef.current!.present();
+    bottomSheetRef.current?.present();
   }, []);
   const handleEnablePanDownToClosePress = useCallback(() => {
     setEnablePanDownToClose(state => !state);