+++ src/components/bottomSheet/styles.ts
@@ -8,10 +8,4 @@ export const styles = StyleSheet.create({
     left: 0,
     right: 0,
   },
-  contentContainer: {
-    overflow: 'visible',
-  },
-  contentMaskContainer: {
-    overflow: 'hidden',
-  },
 });