+++ src/components/bottomSheetBackdropContainer/types.d.ts
@@ -1,6 +1,0 @@
-import type { BottomSheetProps } from '../bottomSheet';
-import type { BottomSheetBackdropProps } from '../bottomSheetBackdrop';
-
-export interface BottomSheetBackdropContainerProps
-  extends Pick<BottomSheetProps, 'backdropComponent'>,
-    BottomSheetBackdropProps {}