+++ src/components/bottomSheetBackground/index.ts
@@ -1,2 +1,2 @@
-export { default } from './BottomSheetBackground';
+export { BottomSheetBackgroundContainer } from './BottomSheetBackgroundContainer';
 export type { BottomSheetBackgroundProps } from './types';