+++ src/components/bottomSheetFooter/index.ts
@@ -1,2 +1,3 @@
-export { default } from './BottomSheetFooter';
+export { BottomSheetFooter } from './BottomSheetFooter';
+export { BottomSheetFooterContainer } from './BottomSheetFooterContainer';
 export type { BottomSheetFooterProps } from './types';