+++ src/hooks/useScrollableSetter.ts
@@ -1,9 +1,9 @@
 import type React from 'react';
 import { useCallback, useEffect } from 'react';
-import { findNodeHandle } from 'react-native';
 import type { SharedValue } from 'react-native-reanimated';
 import type { SCROLLABLE_TYPE } from '../constants';
 import type { Scrollable } from '../types';
+import { findNodeHandle } from '../utilities';
 import { useBottomSheetInternal } from './useBottomSheetInternal';
 
 export const useScrollableSetter = (