+++ src/index.ts
@@ -27,13 +27,15 @@ export {
   BottomSheetVirtualizedList,
   BottomSheetFlashList,
 } from './components/bottomSheetScrollable';
-export { default as BottomSheetHandle } from './components/bottomSheetHandle';
+export { BottomSheetHandle } from './components/bottomSheetHandle';
 export { default as BottomSheetDraggableView } from './components/bottomSheetDraggableView';
 export { default as BottomSheetView } from './components/bottomSheetView';
 export { default as BottomSheetTextInput } from './components/bottomSheetTextInput';
-export { default as BottomSheetBackdrop } from './components/bottomSheetBackdrop';
-export { default as BottomSheetFooter } from './components/bottomSheetFooter';
-export { default as BottomSheetFooterContainer } from './components/bottomSheetFooterContainer/BottomSheetFooterContainer';
+export { BottomSheetBackdrop } from './components/bottomSheetBackdrop';
+export {
+  BottomSheetFooter,
+  BottomSheetFooterContainer,
+} from './components/bottomSheetFooter';
 
 // touchables
 import BottomSheetTouchable from './components/touchables';