+++ src/utilities/index.ts
@@ -3,3 +3,5 @@ export { animate } from './animate';
 export { getKeyboardAnimationConfigs } from './getKeyboardAnimationConfigs';
 export { print } from './logger';
 export { noop, workletNoop } from './noop';
+export { isFabricInstalled } from './isFabricInstalled';
+export { findNodeHandle } from './findNodeHandle';